+++ lib/commanded/process_managers/process_router.ex
@@ -6,6 +6,7 @@ defmodule Commanded.ProcessManagers.ProcessRouter do
 
   require Logger
 
+  alias Commanded.Event.Upcast
   alias Commanded.EventStore
   alias Commanded.EventStore.RecordedEvent
   alias Commanded.ProcessManagers.FailureContext
@@ -51,6 +52,7 @@ defmodule Commanded.ProcessManagers.ProcessRouter do
   end
 
   def init(%State{} = state) do
+    :ok = register_subscription(state)
     :ok = GenServer.cast(self(), :subscribe_to_events)
 
     {:ok, state}
@@ -154,14 +156,7 @@ defmodule Commanded.ProcessManagers.ProcessRouter do
   def handle_info({:subscribed, subscription}, %State{subscription: subscription} = state) do
     Logger.debug(fn -> describe(state) <> " has successfully subscribed to event store" end)
 
-    %State{
-      consistency: consistency,
-      command_dispatcher: command_dispatcher,
-      process_manager_name: process_manager_name
-    } = state
-
-    # Register this process manager as a subscription with the given consistency
-    :ok = Subscriptions.register(process_manager_name, consistency)
+    %State{command_dispatcher: command_dispatcher} = state
 
     {:ok, supervisor} = Supervisor.start_link(command_dispatcher, self())
 
@@ -174,7 +169,10 @@ defmodule Commanded.ProcessManagers.ProcessRouter do
 
     %State{pending_events: pending_events} = state
 
-    unseen_events = Enum.reject(events, &event_already_seen?(&1, state))
+    unseen_events =
+      events
+      |> Enum.reject(&event_already_seen?(&1, state))
+      |> Upcast.upcast_event_stream()
 
     state =
       case {pending_events, unseen_events} do
@@ -244,6 +242,13 @@ defmodule Commanded.ProcessManagers.ProcessRouter do
     Logger.warn(fn -> describe(state) <> " is stopping due to: #{inspect(reason)}" end)
 
     {:stop, reason, state}
+  end
+
+  # Register this process manager as a subscription with the given consistency
+  defp register_subscription(%State{} = state) do
+    %State{consistency: consistency, process_manager_name: name} = state
+
+    Subscriptions.register(name, consistency)
   end
 
   defp subscribe_to_all_streams(%State{} = state) do