+++ CHANGELOG.md
@@ -5,7 +5,13 @@
 ### Enhancements
 
 - Update typespec for `data` and `metadata` fields in `Commanded.EventStore.EventData` struct ([#246](https://github.com/commanded/commanded/pull/246)).
+- Add `include_execution_result` and `aggregate_version` to typespec for router command dispatch ([#262](https://github.com/commanded/commanded/pull/262)).
 - Add `.formatter.exs` to Hex package ([#247](https://github.com/commanded/commanded/pull/247)).
+- Event upcasting ([#263](https://github.com/commanded/commanded/pull/263)).
+- Support `:ok` tagged tuple events from aggregate ([#268](https://github.com/commanded/commanded/pull/268)).
+- Modify `Commanded.Registration.start_child/3` to pass a child_spec ([#273](https://github.com/commanded/commanded/pull/273)).
+- Add `supervisor_child_spec/2` to `Commanded.Registration` behaviour ([#277](https://github.com/commanded/commanded/pull/277)) used by [Commanded Horde Registry](https://github.com/uberbrodt/commanded_horde_registry).
+- Ensure Commanded can be compiled when optional Jason dependency is not present ([#286](https://github.com/commanded/commanded/pull/286)).
 
 ## v0.18.0
 
@@ -105,6 +111,12 @@ Commanded Ecto projections:
 Commanded scheduler:
 
 - `commanded_scheduler` to [v0.2.0](https://hex.pm/packages/commanded_scheduler)
+
+## v0.17.4
+
+### Bug fixes
+
+- Register event handler and process manager subscriptions on process start ([#272](https://github.com/commanded/commanded/pull/272)).
 
 ## v0.17.3